Remove product.models.utils file
