+++ tests/test_product.py
@@ -10,8 +10,7 @@ from django.utils.encoding import smart_text
 from saleor.cart.models import Cart
 from saleor.cart import CartStatus, utils
 from saleor.product import models
-from saleor.product.models.utils import get_attributes_display_map
-from saleor.product.utils import get_availability
+from saleor.product.utils import get_availability, get_attributes_display_map
 from tests.utils import filter_products_by_attribute
